better cobblegen, many new commands, bug fixes, reformatting
The cobblestone generator has been improved greatly; it is made in the same way a vanilla cobblegen is made.

Reformatted some of the old code and edited the help menu.

New information in the /is info <player> command

Fixed a /is ban bug.

Added an island ranking system and some commands correlating to it.

Added an island leaderboard to view the top islands on a server.

New option in the config to change the initial island boundaries created upon generating an island.

New commands to increase or decrease the island boundaries of players.

new commands:
/is rank: Use this command to calculate your island rank on a server.
/is top: Use this command to view the top 5 islands on a server.
/is increase <amount> <player>: Use this command to increase the island boundaries of a player.
/is decrease <amount> <player>: Use this command to decrease the island boundaries of a player.